Some useful experiments; a lot needs to be fixed
I misunderstood Kevtris's doc; the display clock uses an inverted clock,
but that doesn't mean the ppu is ticked on the falling edges. This means
all the +1 stuff should be removed from the ppu, I think.

Also, I wasn't able to pass the ly00_* tests, the timing never worked
out properly w/ gekkio's internal ppu data. In particular, there is a
brief period directly after vblank where the CPU can read mode 0, before
it transitions to mode 2.

He took a look at it; it seems what's happening is that the "dummy mode
3" is being written at the same time that the CPU is reading, which
causes the CPU to read 0 for those "live" bits. The unused/read-only
bits seem to stay the same. This causes $82 to be read as $80. His guess
is that the interrupt doesn't fire, as this is not actually mode 0 (but
it's not clear whether that's true)